Move data validation into src as a data stage
